Adds generic SetupRuntime<T> method.
